Add files via upload
Build 1 = Initial build that connected this pipeline to the repo and had the test pass for all cases 
Build 2 = My initial implementation of the new feature. It had failed due to the logic in the function attempting to subtract strings 
Build 3 = Revised function that addresses the issue that came up in Build 2.